// packages/extensions/Domain/DataManager/_JgJpnDataManager.d.ts
export type FieldOfCumulative =
  | '累計'
  | '退院数累計'
  | '死者合計'
  | 'PCR検査実施人数'
export type FieldOfDate = '確定日' | '発症日' | '確定日YYYYMMDD'
export type FieldOfDateForCumulative = '確定日' | '確定日YYYYMMDD'
export type Field =
  | '通し'
  | '厚労省NO'
  | '無症状病原体保有者'
  | '国内'
  | 'チャーター便'
  | '年代'
  | '性別'
  | '確定日'
  | '発症日'
  | '受診都道府県'
  | '居住都道府県'
  | '居住管内'
  | '居住市区町村'
  | 'キー'
  | '発表'
  | '都道府県内症例番号'
  | '市町村内症例番号'
  | 'ステータス'
  | '備考'
  | 'ソース'
  | 'ソース2'
  | 'ソース3'
  | '人数'
  | '累計'
  | '前日比'
  | '発症数'
  | '死者合計'
  | '退院数累計'
  | '退院数'
  | 'PCR検査実施人数'
  | 'PCR検査前日比'
  | '職業_正誤確認用'
  | '勤務先_正誤確認用'
  | 'Hospital Pref'
  | 'Residential Pref'
  | 'Release'
  | 'Gender'
  | 'X'
  | 'Y'
  | '確定日YYYYMMDD'
  | '受診都道府県コード'
  | '居住都道府県コード'
  | '更新日時'
  | 'Field2'
  | 'Field4'
  | 'Field5'
  | 'Field6'
  | 'Field7'
  | 'Field8'
  | 'Field9'
  | 'Field10'